requirements.txt
RPi.GPIO
spidev
suntime
flask
flask-appconfig
flask-bootstrap
flask-nav
flask-debug
flask-wtf
netifaces
